Reduced data leakage and fixed logit predict error.
Instead of using two non-replacement samples, instead used logical vector
of indicies to select. This allows the indicies to be easily inverted and
thus the data to be 'split' instead of randomly selected. Reduces data
leakage from overlapping elements in each set.

Fixed a stupid logit error where I was predicting off the fixed values of
the regression not on the fucking testing data and was wondering why it
wasn't accurate.